+++ src/utils/filterUtils.js
@@ -79,3 +79,57 @@ export const languagesMatchFilter = (lessonLanguages, languageFilter) => {
   const checkedLanguages = getKeysWithTrueValues(languageFilter); // only the checked tags; e.g. ['nb']
   return checkedLanguages.some(checkedTagKey => lessonLanguages.includes(checkedTagKey));
 };
+
+/**
+ * Make sure tagItems is an array. Try to convert to array if it is not.
+ * This happens if tags is created as string or numbers (e.g. someGroupKey: tag1, tag2, 12345)
+ * instead of list (e.g. someGroupKey: [tag1, tag2, 12345]) in YAML
+ * @param tagItems
+ * @returns {Array}
+ */
+export const fixNonArrayTagList = (tagItems) => {
+  if (tagItems == null) return [];
+  if (typeof tagItems === 'string') return tagItems.split(/,\s*/);
+  if (!Array.isArray(tagItems) && typeof tagItems !== 'string') return fixNonArrayTagList(tagItems.toString());
+  return tagItems;
+};
+
+/**
+ * Fix invalid tags
+ * @param {Object} tags
+ * @param {string} src
+ * @returns {Object} valid tags
+ */
+export const cleanseTags = (tags, src) => {
+  if (tags == null) return {};
+
+  const filterkeys = getFilterkeys();
+
+  return Object.keys(tags).reduce((result, groupKey) => {
+    const groupKeyLC = groupKey.toLowerCase();
+    if (!Object.keys(filterkeys).includes(groupKeyLC)) {
+      console.warn('Ignoring invalid group ' + groupKey + ' in ' + src);
+      return result;
+    }
+
+    let tagsInGroup = fixNonArrayTagList(tags[groupKey]).filter( (tagKey) => {
+      const isValid = tagKey.length > 0 && filterkeys[groupKeyLC].includes(tagKey.toLowerCase());
+      if (!isValid) {
+        console.warn('Ignoring invalid tag ' + tagKey + ' in group ' + groupKey + ' in ' + src);
+      }
+      return isValid;
+    });
+
+    if (tagsInGroup.length === 0) { return result; } // Ignore tagGroups with no tagItems
+
+    result[groupKey.toLowerCase()] = tagsInGroup.map(tagKey => tagKey.toLowerCase());
+    return result;
+  }, {});
+};
+
+/**
+* Returns languages defined as available
+* All available languages must be defined here
+* @returns {string[]} An array of available languages, e.g. ['nb', 'nn', 'en']
+*/
+export const getAvailableLanguages = () => getFilterkeys().language;
